[DAGE-25] Moved factory for llm and adapted code in dataset_generator.py
