+++ rre-dataset-generator/src/llm/llm_provider_factory.py
@@ -44,18 +44,3 @@ def build_gemini(config: LLMConfig) -> BaseChatModel:
         max_output_tokens=config.max_tokens,
         google_api_key=key,
     )
-
-
-PROVIDER_REGISTRY = {
-    "openai": build_openai,
-    "gemini": build_gemini,
-}
-
-
-def build_chat_model(config: LLMConfig) -> BaseChatModel:
-    provider_name = config.name
-    if provider_name not in PROVIDER_REGISTRY:
-        log.error("Unsupported LLM provider requested: %s", provider_name)
-        raise ValueError(f"Unsupported provider: {provider_name}")
-    log.info("Selected LLM provider: %s", provider_name)
-    return PROVIDER_REGISTRY[provider_name](config)
